refactor(api): extract obsidian parsers to separate module
Extracts getExtension, isImageFile, isMarkdownFile, and
parseFrontmatter from obsidian-import.ts into a dedicated
obsidian-parsers.ts module for unit testability.

The existing unit test now imports from the shared module
instead of duplicating parser internals.
